refactor(bench): split into v0.4.0-swept (benchmarks) and HEAD-baseline (baseline_head, projection) targets with version-floor annotations
