+++ benches/benchmarks.rs
@@ -2,10 +2,11 @@
 //!
 //! Run with: cargo bench
 //! Run specific benchmark: cargo bench -- parsing
+//!
+//! VERSION FLOOR: v0.4.0 — SWEPT target. Every group/case here parses and normalizes at v0.4.0 (verified by probe), so scripts/perf/sweep_tags.sh overlays this file onto all release tags for regression detection. Do NOT add cases that don't compile+resolve at v0.4.0 — put those in baseline_head.rs.
 
 use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
 use ferro_hgvs::hgvs::parser::parse_hgvs_fast;
-use ferro_hgvs::reference::{Exon, Strand, Transcript};
 use ferro_hgvs::{parse_hgvs, MockProvider, Normalizer};
 use flate2::read::GzDecoder;
 use serde::Deserialize;
@@ -53,7 +54,6 @@ fn bench_parsing(c: &mut Criterion) {
         // --- coverage gap-fills (churned grammar) ---
         ("m.sub", "NC_012920.1:m.8993T>G"),
         ("allele_cis", "NC_000001.11:g.[100A>G;200del]"),
-        ("allele_trans", "NC_000001.11:g.[100A>G];[200del]"),
         ("c.compound_repeat", "NM_000088.3:c.5GA[10]"),
         ("c.uncertain", "NM_000088.3:c.(100_200)del"),
         ("p.predicted", "NP_000079.2:p.(Val600Glu)"),
@@ -179,78 +179,9 @@ fn bench_parsing_throughput(c: &mut Criterion) {
 // Normalization benchmarks
 // =============================================================================
 
-/// `with_test_data()` plus a genome-mapped contig and two intron-bearing
-/// transcripts (one per strand), so `g.` and intronic normalization run on
-/// the real path. Uses only reference APIs present since v0.4.0, so this
-/// file overlays onto every release tag during the sweep.
-fn enriched_provider() -> MockProvider {
-    use ferro_hgvs::reference::transcript::{GenomeBuild, ManeStatus};
-    let mut p = MockProvider::with_test_data();
-
-    // Plus-strand, two-exon coding transcript with a real intron, mapped to
-    // NC_000001.11. Exon1 genome[1000,1009] tx[1,10]; intron genome[1010,1999];
-    // exon2 genome[2000,2009] tx[11,20]. CDS tx[1,18]; tx[19,20] is a 2-nt 3'UTR tail.
-    p.add_transcript(Transcript::new(
-        "NM_INTR.1".to_string(),
-        Some("INTRGENE".to_string()),
-        Strand::Plus,
-        Some("ATGCGCAAAGGGTAACCCNN".to_string()), // 20 bp (18 bp CDS + 2 nt 3'UTR tail)
-        Some(1),
-        Some(18),
-        vec![
-            Exon::with_genomic(1, 1, 10, 1000, 1009),
-            Exon::with_genomic(2, 11, 20, 2000, 2009),
-        ],
-        Some("NC_000001.11".to_string()),
-        Some(1000),
-        Some(2009),
-        GenomeBuild::default(),
-        ManeStatus::default(),
-        None,
-        None,
-    ));
-
-    // Minus-strand counterpart for the #497 minus-strand intronic shuffle path.
-    // On minus strand exons are listed 5'→3' in transcript coordinates, but
-    // their genomic coords are reversed: exon1 (tx 1..10) maps to genome 2000..2009,
-    // exon2 (tx 11..20) maps to genome 1000..1009.
-    p.add_transcript(Transcript::new(
-        "NM_INTRM.1".to_string(),
-        Some("INTRGENM".to_string()),
-        Strand::Minus,
-        Some("ATGCGCAAAGGGTAACCCNN".to_string()), // 20 bp (18 bp CDS + 2 nt 3'UTR tail)
-        Some(1),
-        Some(18),
-        vec![
-            Exon::with_genomic(1, 1, 10, 2000, 2009),
-            Exon::with_genomic(2, 11, 20, 1000, 1009),
-        ],
-        Some("NC_000001.11".to_string()),
-        Some(1000),
-        Some(2009),
-        GenomeBuild::default(),
-        ManeStatus::default(),
-        None,
-        None,
-    ));
-
-    // Genomic sequence for NC_000001.11: 999 N's + exon1 + 990 N intron +
-    // exon2 + 100 N's, so positions 1000..2009 carry the modeled bases.
-    let prefix = "N".repeat(999);
-    let exon1 = "ATGCGCAAAG"; // genome 1000..1009
-    let intron = "N".repeat(990); // genome 1010..1999
-    let exon2 = "GGTAACCCNN"; // genome 2000..2009
-    let suffix = "N".repeat(100);
-    p.add_genomic_sequence(
-        "NC_000001.11",
-        format!("{prefix}{exon1}{intron}{exon2}{suffix}"),
-    );
-    p
-}
-
 /// Benchmark normalization performance for different variant types
 fn bench_normalization(c: &mut Criterion) {
-    let provider = enriched_provider();
+    let provider = MockProvider::with_test_data();
     let normalizer = Normalizer::new(provider);
 
     let test_cases = vec![
@@ -267,20 +198,11 @@ fn bench_normalization(c: &mut Criterion) {
         ("c.sub_minus", "NM_999999.1:c.5A>G"),
         ("c.del_minus", "NM_999999.1:c.5del"),
         ("c.ins_minus", "NM_999999.1:c.5_6insAT"),
-        // c. intronic (enriched fixture; #497 shuffle path)
-        ("c.intronic_plus_del", "NM_INTR.1:c.10+2del"),
-        ("c.intronic_minus_del", "NM_INTRM.1:c.10+2del"),
         // protein (NP_000079.2: M1, R97, V600)
         ("p.sub", "NP_000079.2:p.Val600Glu"),
         ("p.del", "NP_000079.2:p.Val600del"),
         ("p.fs", "NP_000079.2:p.Arg97fs"),
         ("p.fsTer", "NP_000079.2:p.Arg97fsTer15"),
-        // g. (enriched fixture, NC_000001.11)
-        ("g.sub", "NC_000001.11:g.1000A>G"),
-        ("g.del", "NC_000001.11:g.1001del"),
-        ("g.ins", "NC_000001.11:g.1001_1002insTT"),
-        ("g.delins", "NC_000001.11:g.1001_1003delinsTT"),
-        ("g.inv", "NC_000001.11:g.1001_1003inv"),
     ];
 
     // Guard: every case must parse AND normalize on the real path. A case
@@ -558,28 +480,6 @@ fn bench_allele_scaling(c: &mut Criterion) {
     group.finish();
 }
 
-/// Parse-path timing for inputs that exercise the W30xx correction/diagnostic
-/// machinery (swapped positions, dup/del soft-prohibition suffix forms).
-fn bench_correctors(c: &mut Criterion) {
-    // These inputs exercise the W30xx correction/diagnostic machinery:
-    //   swapped_positions (W3026): Ok-with-warning in WarnCorrect mode, Err in Reject mode.
-    //   dup_count_suffix / del_seq_suffix / dup_seq_suffix: Ok-with-warning under the lenient default.
-    // We time the full parse path regardless of Ok/Err — that is intentional, so there is no is_ok guard.
-    let cases = vec![
-        ("swapped_positions", "NC_000001.11:g.200_100del"),
-        ("dup_count_suffix", "NM_000088.3:c.100_102dup3"),
-        ("del_seq_suffix", "NM_000088.3:c.100delA"),
-        ("dup_seq_suffix", "NM_000088.3:c.100_102dupATG"),
-    ];
-    let mut group = c.benchmark_group("correctors");
-    for (name, input) in &cases {
-        group.bench_with_input(BenchmarkId::new("parse", name), input, |b, v| {
-            b.iter(|| parse_hgvs(black_box(v)))
-        });
-    }
-    group.finish();
-}
-
 criterion_group!(
     benches,
     bench_parsing,
@@ -592,7 +492,6 @@ criterion_group!(
     bench_fast_path,
     bench_error_handling,
     bench_allele_scaling,
-    bench_correctors,
 );
 
 criterion_main!(benches);
